requirements.txt
astroid==1.4.9
boto==2.42.0
dj-database-url==0.4.2
Django==1.10
django-autocomplete-light==3.1.8
django-cors-headers==2.0.2
django-filter==1.0.1
django-grappelli==2.8.1
django-money==0.10.2
django-storages==1.4.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
drfdocs==0.0.11
gunicorn==19.7.1
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.8
mccabe==0.6.1
Pillow==3.3.0
psycopg2==2.6.2
py-moneyed==0.6.0
PyJWT==1.4.2
pylint==1.6.5
pytz==2016.6.1
PyYAML==3.12
requests==2.11.1
six==1.10.0
whitenoise==3.3.0
wrapt==1.10.10
